+++ setup.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+
+echo Creating symlinks for TarPacker...
+cwd = pwd
+ln -s $cwd/tp.py /usr/bin/tp